pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["ibraheem-111 <ibraheemsal@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
python-dotenv = "^1.0.1"
psycopg = "^3.2.1"
langchain-postgres = "^0.0.9"
langchain-openai = "^0.1.17"
psycopg-binary = "^3.2.1"
langchain = "^0.2.11"
firebase-admin = "^6.5.0"
pypdf = "^4.3.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.2"}
aioredis = "^2.0.1"
setuptools = "^71.1.0"
boto3 = "^1.34.151"
langchain-aws = "^0.1.16"
pymupdf4llm = {git = "https://github.com/ibraheem-111/RAG.git", subdirectory = "pymupdf4llm"}
exceptiongroup = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
